Make it possible to dry run planned migrations
All of these are post-merge so we can use pre-merge flag to just log statements without doing the actual deletions yet.

+++ src/server/datastore/prompts.ts
@@ -30,9 +30,6 @@ import { PropertyFilter, and } from '@google-cloud/datastore'
 import { deleteEntity } from './cleanup'
 
 export async function migratePrompts(postMerge: boolean) {
-  if (!postMerge) {
-    return
-  }
   const datastore = getDatastore()
   const [allPrompts] = await datastore.runQuery(datastore.createQuery(Entity.PROMPT))
   const usedProjectIDs = new Set(allPrompts.map(promptData => promptData.projectID))
@@ -40,9 +37,11 @@ export async function migratePrompts(postMerge: boolean) {
   const allProjectIDs = new Set(allProjects.map(project => getID(project)))
   console.log(`Found ${allPrompts.length} prompts (for ${usedProjectIDs.size} projects out of ${allProjectIDs.size})`)
   for (const promptData of allPrompts) {
-    if (!allProjectIDs.has(promptData.projectID)) {
+    if (!!promptData.projectID && !allProjectIDs.has(promptData.projectID)) {
       console.log(`Deleting prompt ${getID(promptData)} for missing project ${promptData.projectID}`)
-      await datastore.delete(buildKey(Entity.PROMPT, getID(promptData)))
+      if (postMerge) {
+        await datastore.delete(buildKey(Entity.PROMPT, getID(promptData)))
+      }
     }
     // await updatePrompt({ ...promptData }, false)
   }
